move SetDefaultHTTP to middleware
