+++ server/http.go
@@ -3,14 +3,12 @@ package server
 import (
 	"context"
 	"fmt"
-	"net/http"
 
 	"github.com/labstack/echo-contrib/prometheus"
 	"github.com/labstack/echo/v4"
 	echoMW "github.com/labstack/echo/v4/middleware"
 	"github.com/labstack/gommon/log"
 	"github.com/webdevelop-pro/go-common/configurator"
-	"github.com/webdevelop-pro/go-common/context/keys"
 	"github.com/webdevelop-pro/go-common/server/healthcheck"
 	"github.com/webdevelop-pro/go-common/server/middleware"
 	"github.com/webdevelop-pro/go-common/server/validator"
@@ -81,6 +79,7 @@ func NewHttpServer(e *echo.Echo, l logger.Logger, cfg *Config, authTool middlewa
 			c.Set(echo.HeaderXRequestID, requestID)
 		},
 	}))
+	e.Use(middleware.DefaultCTXValues)
 	// Add the healthcheck endpoint
 	e.GET(`/healthcheck`, healthcheck.Healthcheck)
 
@@ -144,22 +143,4 @@ func StartServer(lc fx.Lifecycle, srv *HttpServer) {
 			},
 		},
 	)
-}
-
-func SetDefaultHTTP(echoContext echo.Context) context.Context {
-	ctx := SetDefaultHTTPCtx(echoContext.Request().Context(), echoContext.Request().Header)
-	if ctx.Value(keys.RequestID).(string) == "" {
-		requestID := echoContext.Get(echo.HeaderXRequestID).(string)
-		ctx = keys.SetCtxValue(ctx, keys.RequestID, requestID)
-	}
-	return ctx
-}
-
-func SetDefaultHTTPCtx(ctx context.Context, headers http.Header) context.Context {
-	requestID := headers.Get(echo.HeaderXRequestID)
-	IP := middleware.GetIpAddress(headers)
-
-	ctx = keys.SetCtxValue(ctx, keys.RequestID, requestID)
-	ctx = keys.SetCtxValue(ctx, keys.IPAddress, IP)
-	return ctx
 }
