+++ server/http_test.go
@@ -10,6 +10,7 @@ import (
 	"github.com/labstack/echo/v4"
 	"github.com/stretchr/testify/assert"
 	"github.com/webdevelop-pro/go-common/context/keys"
+	"github.com/webdevelop-pro/go-common/server/middleware"
 )
 
 func TestHTTPCtx(t *testing.T) {
@@ -19,7 +20,7 @@ func TestHTTPCtx(t *testing.T) {
 		"X-Forwarded-For": {"31.6.1.12"},
 	}
 
-	ctx = SetDefaultHTTPCtx(ctx, headers)
+	ctx = middleware.SetDefaultHTTPCtx(ctx, headers)
 
 	assert.Equal(t, headers["X-Request-Id"][0], keys.GetCtxValue(ctx, keys.RequestID))
 	assert.Equal(t, headers["X-Forwarded-For"][0], keys.GetCtxValue(ctx, keys.IPAddress))
@@ -33,6 +34,6 @@ func TestEmptyRequestID(t *testing.T) {
 	rec := httptest.NewRecorder()
 	echoCtx := e.NewContext(req, rec)
 	echoCtx.Set(echo.HeaderXRequestID, "123123123")
-	ctx := SetDefaultHTTP(echoCtx)
+	ctx := middleware.SetDefaultCTX(echoCtx)
 	assert.Equal(t, len(keys.GetCtxValue(ctx, keys.RequestID).(string)), 9)
 }
